refactor: Path 관련 파일들 적절한 package로 이동
