+++ src/main/java/nextstep/subway/path/application/PathService.java
@@ -1,10 +1,9 @@
-package nextstep.subway.line.application;
+package nextstep.subway.path.application;
 
 import nextstep.subway.auth.domain.LoginMember;
-import nextstep.subway.line.domain.Line;
-import nextstep.subway.line.domain.PathFinder;
-import nextstep.subway.line.domain.Section;
-import nextstep.subway.line.dto.PathResponse;
+import nextstep.subway.line.application.LineService;
+import nextstep.subway.path.domain.PathFinder;
+import nextstep.subway.path.dto.PathResponse;
 import nextstep.subway.station.application.StationService;
 import nextstep.subway.station.domain.Station;
 import org.springframework.stereotype.Service;
