+++ src/test/java/study/unit/MockitoExtensionTest.java
@@ -4,7 +4,7 @@ import com.google.common.collect.Lists;
 import nextstep.subway.line.application.LineService;
 import nextstep.subway.line.domain.Line;
 import nextstep.subway.line.domain.LineRepository;
-import nextstep.subway.line.domain.PathFinder;
+import nextstep.subway.path.domain.PathFinder;
 import nextstep.subway.line.dto.LineResponse;
 import nextstep.subway.station.application.StationService;
 import org.junit.jupiter.api.DisplayName;
